+++ packages/player/src/components/core/clickable-component.tsx
