+++ packages/player/src/components/core/slider.tsx
