+++ packages/player/src/components/cue-bar.tsx
@@ -4,12 +4,12 @@ import {isEmpty} from 'lodash'
 import Tippy from '@tippyjs/react'
 import {scroller} from 'react-scroll'
 import {track} from '@skillrecordings/analytics'
-import {useNotesCues} from '../../hooks/use-notes-cues'
+import {useNotesCues} from '../hooks/use-notes-cues'
 import CodeBlock from '@skillrecordings/react/dist/components/code-block'
-import {useVideo} from '../../context/video-context'
+import {useVideo} from '../context/video-context'
 import {useSelector} from '@xstate/react'
-import {selectDuration, selectVideo} from '../../selectors'
-import {useCue} from '../../hooks/use-cue'
+import {selectDuration, selectVideo} from '../selectors'
+import {useCue} from '../hooks/use-cue'
 import {ElementType} from 'react'
 import ReactMarkdown from 'react-markdown'
