+++ packages/player/src/components/player.tsx
@@ -9,7 +9,7 @@ import {LoadingSpinner} from './loading-spinner'
 import {BigPlayButton} from './big-play-button'
 import {Bezel} from './bezel'
 import {ControlBar} from './control-bar'
-import {CueBar} from './controls/cue-bar'
+import {CueBar} from './cue-bar'
 
 import * as browser from '../utils/browser'
 
@@ -22,6 +22,7 @@ import {
   selectIsWaiting,
   selectVideo,
 } from '../selectors'
+import {ProgressBar} from './progress-bar'
 
 type PlayerProps = {
   container?: HTMLElement
@@ -189,6 +190,7 @@ export const Player: React.FC<PlayerProps> = (props) => {
         <Bezel />
         <LoadingSpinner />
       </div>
+      <ProgressBar />
       <CueBar />
       <ControlBar />
       <Shortcut />
