+++ packages/player/src/components/seek-bar/play-progress-bar.tsx
