+++ packages/player/src/index.ts
@@ -1,5 +1,5 @@
 export {Player} from './components/player'
 export {VideoProvider} from './context/video-context'
 export {HLSSource} from './components/hls-source'
-export {CueBar} from './components/controls/cue-bar'
+export {CueBar} from './components/cue-bar'
 export {useNotesCues} from './hooks/use-notes-cues'
