+++ packages/player/src/styles/scss/index.scss
@@ -15,6 +15,7 @@
 @import 'components/slider';
 @import 'components/control';
 @import 'components/control-bar';
+@import 'components/progress-bar';
 @import 'components/progress';
 @import 'components/play-pause';
 @import 'components/fullscreen';
